Gather sources
