+++ feature/CourseView/Sources/CourseView/ExerciseTab/ExerciseListView.swift
@@ -140,7 +140,7 @@ struct ExerciseListCell: View {
             } else {
                 Text(R.string.localizable.noDueDate())
             }
-            SubmissionResultStatusView(exercise: exercise) // a
+            SubmissionResultStatusView(exercise: exercise)
             ScrollView(.horizontal) {
                 LazyHGrid(rows: rows, spacing: .s) {
                     if let releaseDate = exercise.baseExercise.releaseDate,
