+++ feature/Dashboard/Package.swift
@@ -16,16 +16,13 @@ let package = Package(
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
-        .package(url: "https://github.com/ls1intum/artemis-ios-core-modules", branch: "feature/development/screenshots"),
+        .package(path: "../../../artemis-ios-core-modules"),
         .package(path: "../../core/Navigation"),
         .package(path: "../CourseRegistration"),
         .package(path: "../Notifications"),
         .package(path: "../CourseView"),
         .package(url: "https://github.com/mac-cain13/R.swift.git", from: "7.0.0"),
         .package(url: "https://github.com/pointfreeco/swift-dependencies", from: "1.0.0"),
-        .package(url: "https://github.com/pointfreeco/swift-snapshot-testing", from: "1.0.0"),
-        // Fix error in SwiftStomp
-        .package(url: "https://github.com/daltoniam/Starscream.git", exact: "4.0.4"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -45,16 +42,10 @@ let package = Package(
                 .product(name: "RswiftLibrary", package: "R.swift"),
                 .product(name: "Dependencies", package: "swift-dependencies"),
             ],
-            resources: [
-                .copy("Media"),
-            ],
             plugins: [.plugin(name: "RswiftGeneratePublicResources", package: "R.swift")]
         ),
         .testTarget(
             name: "DashboardTests",
-            dependencies: [
-                "Dashboard",
-                .product(name: "SnapshotTesting", package: "swift-snapshot-testing"),
-            ])
+            dependencies: ["Dashboard"])
     ]
 )
