+++ feature/Dashboard/Tests/DashboardTests/DashboardTests.swift
@@ -1,24 +1,10 @@
-import Dependencies
-import SnapshotTesting
-import SwiftUI
 import XCTest
 @testable import Dashboard
 
-@MainActor
 final class DashboardTests: XCTestCase {
-    let record = true
-
-    func testDashboardSnapshot() {
-        let viewModel = withDependencies { values in
-            values.courseService = CourseServiceStub()
-        } operation: {
-            CoursesOverviewViewModel()
-        }
-        let view = NavigationStack {
-            CoursesOverviewView(viewModel: viewModel)
-        }
-        assertSnapshot(of: view,
-                       as: .wait(for: 10, on: .image(layout: .device(config: .iPhone13ProMax))),
-                       record: record)
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
     }
 }
