+++ feature/Screenshots/Sources/Screenshots/CourseServiceStub.swift
