+++ feature/Screenshots/Sources/Screenshots/CourseView+Preview.swift
@@ -11,6 +11,7 @@ import Navigation
 import SharedModels
 import SharedServices
 import SwiftUI
+@testable import CourseView
 
 struct CoursesView_Previews: PreviewProvider {
